+++ backoff/backoff_constant.go
@@ -1,26 +1,29 @@
-package retry
+package backoff
 
 import (
 	"context"
 	"fmt"
 	"time"
+
+	"github.com/swayne275/go-retry"
+	"github.com/swayne275/go-retry/common/backoff"
 )
 
 // Constant is a wrapper around Retry that uses a constant backoff. It will
 // retry the function f until it returns an error, or the context is canceled.
 // TODO is this really useful vs an example? would have to extend with a repeat version too.
-func Constant(ctx context.Context, t time.Duration, f RetryFunc) error {
+func Constant(ctx context.Context, t time.Duration, f retry.RetryFunc) error {
 	b, err := NewConstant(t)
 	if err != nil {
 		return fmt.Errorf("failed to create constant backoff: %w", err)
 	}
 
-	return Do(ctx, b, f)
+	return retry.Do(ctx, b, f)
 }
 
 // NewConstant creates a new constant backoff using the value t. The wait time
 // is the provided constant value. It returns an error if t is not greater than 0.
-func NewConstant(t time.Duration) (Backoff, error) {
+func NewConstant(t time.Duration) (backoff.Backoff, error) {
 	if t <= 0 {
 		return nil, fmt.Errorf("constant backoff must be greater than zero")
 	}
