+++ backoff/backoff_exponential_test.go
@@ -1,4 +1,4 @@
-package retry_test
+package backoff_test
 
 import (
 	"math"
@@ -7,10 +7,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/swayne275/go-retry"
+	"github.com/swayne275/go-retry/backoff"
+	cb "github.com/swayne275/go-retry/common/backoff"
 )
 
-func TestFibonacciBackoff(t *testing.T) {
+func TestExponentialBackoff(t *testing.T) {
 	t.Parallel()
 
 	cases := []struct {
@@ -29,36 +30,24 @@ func TestFibonacciBackoff(t *testing.T) {
 			},
 		},
 		{
-			name:  "max",
-			base:  10 * time.Millisecond,
-			tries: 5,
-			exp: []time.Duration{
-				10 * time.Millisecond,
-				20 * time.Millisecond,
-				30 * time.Millisecond,
-				50 * time.Millisecond,
-				80 * time.Millisecond,
-			},
-		},
-		{
 			name:  "many",
 			base:  1 * time.Nanosecond,
 			tries: 14,
 			exp: []time.Duration{
 				1 * time.Nanosecond,
 				2 * time.Nanosecond,
-				3 * time.Nanosecond,
-				5 * time.Nanosecond,
+				4 * time.Nanosecond,
 				8 * time.Nanosecond,
-				13 * time.Nanosecond,
-				21 * time.Nanosecond,
-				34 * time.Nanosecond,
-				55 * time.Nanosecond,
-				89 * time.Nanosecond,
-				144 * time.Nanosecond,
-				233 * time.Nanosecond,
-				377 * time.Nanosecond,
-				610 * time.Nanosecond,
+				16 * time.Nanosecond,
+				32 * time.Nanosecond,
+				64 * time.Nanosecond,
+				128 * time.Nanosecond,
+				256 * time.Nanosecond,
+				512 * time.Nanosecond,
+				1024 * time.Nanosecond,
+				2048 * time.Nanosecond,
+				4096 * time.Nanosecond,
+				8192 * time.Nanosecond,
 			},
 		},
 		{
@@ -68,11 +57,11 @@ func TestFibonacciBackoff(t *testing.T) {
 			exp: []time.Duration{
 				100_000 * time.Hour,
 				200_000 * time.Hour,
-				300_000 * time.Hour,
-				500_000 * time.Hour,
+				400_000 * time.Hour,
 				800_000 * time.Hour,
-				1_300_000 * time.Hour,
-				2_100_000 * time.Hour,
+				1_600_000 * time.Hour,
+				math.MaxInt64,
+				math.MaxInt64,
 				math.MaxInt64,
 				math.MaxInt64,
 				math.MaxInt64,
@@ -93,7 +82,7 @@ func TestFibonacciBackoff(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
-			b, err := retry.NewFibonacci(tc.base)
+			b, err := backoff.NewExponential(tc.base)
 			if tc.expectErr && err == nil {
 				t.Fatal("expected an error")
 			}
@@ -129,26 +118,24 @@ func TestFibonacciBackoff(t *testing.T) {
 	}
 }
 
-func TestFibonacciBackoff_WithReset(t *testing.T) {
-	base := 1 * time.Second
-	numRounds := 5
+func TestExponentialBackoff_WithReset(t *testing.T) {
+	base := 2 * time.Second
+	numRounds := 3
 	expected := []time.Duration{
-		1 * time.Second,
 		2 * time.Second,
-		3 * time.Second,
-		5 * time.Second,
+		4 * time.Second,
 		8 * time.Second,
 	}
 
-	b, err := retry.NewFibonacci(base)
+	b, err := backoff.NewExponential(base)
 	if err != nil {
-		t.Fatalf("failed to create fibonacci backoff: %v", err)
+		t.Fatalf("failed to create exponential backoff: %v", err)
 	}
 
-	resettableB := retry.WithReset(func() retry.Backoff {
-		newB, err := retry.NewFibonacci(base)
+	resettableB := backoff.WithReset(func() cb.Backoff {
+		newB, err := backoff.NewExponential(base)
 		if err != nil {
-			t.Fatalf("failed to reset fibonacci backoff: %v", err)
+			t.Fatalf("failed to reset exponential backoff: %v", err)
 		}
 
 		return newB
@@ -172,75 +159,22 @@ func TestFibonacciBackoff_WithReset(t *testing.T) {
 	}
 }
 
-func TestFibonacciBackoff_WithReset_ChangeBase(t *testing.T) {
-	base := 1 * time.Second
-	numRounds := 5
-	expected := []time.Duration{
-		1 * time.Second,
-		2 * time.Second,
-		3 * time.Second,
-		5 * time.Second,
-		8 * time.Second,
-	}
-
-	b, err := retry.NewFibonacci(base)
-	if err != nil {
-		t.Fatalf("failed to create fibonacci backoff: %v", err)
-	}
-
-	newBase := 2 * time.Second
-	newExpected := []time.Duration{
+func TestExponentialBackoff_WithCappedDuration_WithReset(t *testing.T) {
+	base := 2 * time.Second
+	cappedDuration := 4 * time.Second
+	numRounds := 3
+	expectedCapped := []time.Duration{
 		2 * time.Second,
 		4 * time.Second,
-		6 * time.Second,
-		10 * time.Second,
-		16 * time.Second,
-	}
-	resettableB := retry.WithReset(func() retry.Backoff {
-		newB, err := retry.NewFibonacci(newBase)
-		if err != nil {
-			t.Fatalf("failed to reset fibonacci backoff: %v", err)
-		}
-
-		return newB
-	}, b)
-
-	// test pre reset
-	for i := 0; i < numRounds; i++ {
-		val, _ := resettableB.Next()
-		if val != expected[i] {
-			t.Errorf("pre reset: expected %v to be %v", val, expected[i])
-		}
+		4 * time.Second,
 	}
 
-	// test post reset. since we reset with a new base we expect new values
-	resettableB.Reset()
-	for i := 0; i < numRounds; i++ {
-		val, _ := resettableB.Next()
-		if val != newExpected[i] {
-			t.Errorf("post reset: expected %v to be %v", val, expected[i])
-		}
-	}
-}
-
-func TestFibonacciBackoff_WithCappedDuration_WithReset(t *testing.T) {
-	base := 1 * time.Second
-	cappedDuration := 5 * time.Second
-	numRounds := 5
-	expectedCapped := []time.Duration{
-		1 * time.Second,
-		2 * time.Second,
-		3 * time.Second,
-		5 * time.Second,
-		5 * time.Second,
-	}
-
-	b, err := retry.NewFibonacci(base)
+	b, err := backoff.NewExponential(base)
 	if err != nil {
-		t.Fatalf("failed to create fibonacci backoff: %v", err)
+		t.Fatalf("failed to create exponential backoff: %v", err)
 	}
 
-	cappedB := retry.WithCappedDuration(cappedDuration, b)
+	cappedB := backoff.WithCappedDuration(cappedDuration, b)
 
 	// test pre reset
 	for i := 0; i < numRounds; i++ {
@@ -263,18 +197,15 @@ func TestFibonacciBackoff_WithCappedDuration_WithReset(t *testing.T) {
 	// test post user-defined reset.
 	// since we define the reset function without decorators, the decorators should not be observed.
 	expectedAfterExplicitReset := []time.Duration{
-		1 * time.Second,
 		2 * time.Second,
-		3 * time.Second,
-		5 * time.Second,
+		4 * time.Second,
 		8 * time.Second,
 	}
-
-	resettableB := retry.WithReset(func() retry.Backoff {
+	resettableB := backoff.WithReset(func() cb.Backoff {
 		// don't set a cap on the explicit reset
-		newB, err := retry.NewFibonacci(base)
+		newB, err := backoff.NewExponential(base)
 		if err != nil {
-			t.Fatalf("failed to reset fibonacci backoff: %v", err)
+			t.Fatalf("failed to reset exponential backoff: %v", err)
 		}
 
 		return newB
