+++ backoff/backoff_test.go
@@ -1,8 +1,10 @@
-package retry
+package backoff
 
 import (
 	"testing"
 	"time"
+
+	"github.com/swayne275/go-retry/common/backoff"
 )
 
 func TestWithJitter(t *testing.T) {
@@ -129,7 +131,7 @@ func TestWithMaxDuration(t *testing.T) {
 
 func TestResettableBackoff(t *testing.T) {
 	var attempt uint64
-	b := WithReset(func() Backoff {
+	b := WithReset(func() backoff.Backoff {
 		attempt = 0
 
 		return BackoffFunc(func() (time.Duration, bool) {
